Move setup to new separate ix_apps package
